Add alternative solutions
I tried a few ways of writing the code before deciding on the one I
thought was best. I thought there would be a better way than writing
tuples, but if there is, I couldn't find it! I think the state monad
is the way you would do it, but it would make things super complicated

I'm going to delete these immediately, but wanted to have them in the
history for future reference.