Ensure correctness of normalization
If we normalize some code twice, the second normalization shouldn't
introduce any changes

This is a kind of test / defensive programming strategy, mostly to
catch occassions where a new state returned from normalize is then
ignored (instead of being passed to normalize the next time)

This means that errors can happen in production, but at least the code
is honest, and owns up to its errors instead of hiding them, and it
makes it likely that we would find any problems quickly.

3 tests are skipped in this commit, as the output generated from
Elm-Syntax is invalid for Custom Types. This has been fixed in the repo,
but hasn't made it in to the package for some reason.

See https://github.com/stil4m/elm-syntax/issues/117